A C program that prints the size of various types on the computer it is compiled and run on.

+++ 0x00-hello_world/6-size.c
@@ -14,10 +14,10 @@ int main(void)
 	d = sizeof(long long);
 	e = sizeof(float);
 
-	printf("Size of a char: %d byte(s)\n", (unsigned)a);
-	printf("Size of a int: %d byte(s)\n", (unsigned)b);
-	printf("Size of a long int: %d byte(s)\n", (unsigned)c);
-	printf("Size of a long long: %d byte(s)\n", (unsigned)d);
-	printf("Size of a float: %d byte(s)\n", (unsigned)e);
+	printf("Size of a char: %d byte(s)\n", (unsigned long)a);
+	printf("Size of a int: %d byte(s)\n", (unsigned long)b);
+	printf("Size of a long int: %d byte(s)\n", (unsigned long)c);
+	printf("Size of a long long: %d byte(s)\n", (unsigned long)d);
+	printf("Size of a float: %d byte(s)\n", (unsigned long)e);
 	return (0);
 }
